add cholesky! :smile:

+++ src/utils/reader.py
@@ -1,12 +1,10 @@
-import numpy as np
-
 def getMatrixA(file_path):
   matrixA = []
   with open(file_path, 'r') as file:
     lines = file.readlines()
     for line in lines:
       new_line = line.split(' ')
-      new_line = [int(x) for x in new_line]
+      new_line = [float(x) for x in new_line]
       matrixA.append(new_line)
 
   return matrixA
@@ -15,6 +13,6 @@ def getVectorB(file_path):
   vectorB = []
   with open(file_path, 'r') as file:
     lines = file.readlines()
-    vectorB = [int(x) for x in lines]
+    vectorB = [float(x) for x in lines]
 
   return vectorB
